fileio/adm: reverse track-stream reference direction
The direction of this in the model was somewhat arbitrary -- most of the
time we want to traverse from track to stream rather than stream to
track.

The bigger change here is to make either reference establish this
relationship, and to write both references. A warning is issued if
either is missing (approximately).