fileio/adm: implement audioPackFormat HOA params
This required a change to the block format versions of these parameters
to remove defaults.

These parameters can be specified in more than one place, so the
defaults need to be applied in the item selection code (when we can look
at both the pack and block formats), rather than while
parsing/construction as they were previously.

Otherwise specification of non-default normalisation in an
audioPackFormat would appear to be inconsistent with the default value
in the audioBlockFormats.