use new conversion in Cartesian screen scaling
In polar mode, this behaves identically to the existing solution.

In Cartesian mode, positions are converted to polar using the new
warping conversion function, scaled in polar coordinates, then converted
back.

The advantage of this compared to applying conversion to the screen
representations is that it works with larger screens that go past the
M+-030 loudspeakers -- just scaling the x and y positions is not enough
when the allocentric panner is used, because this only affects the
position between these loudspeakers.