fix rounding of times in ADM XML writer
When writing ADM XML, times would be rounded to 5 digits (with the
possibility of floating point error). This can cause files to be
invalid, because the rtimes and durations might not match up.

This change makes the output have exactly the number of required digits,
if it's representable by a decimal. Notes:

- This doesn't pad to 5 digits; it isn't required and just wastes space.

- When a time is not representable by a decimal, a warning is issued.
  This should be solved outside of the ADM writer, because to do it
  properly requires changing the meaning of the ADM data, and requires
  knowledge of the structures. This means either adjusting things which
  generate ADM to make them generate ns times, or a process which rounds
  the rtimes and then sets the durations to the difference.

- This wasn't generally used by anything public.