select_items: populate pack_absoluteDistance
This generalises _get_extra_data to work with in multi-channel contexts.

For a single channel absoluteDistance can be specified in any of the
audioPackFormats referencing it, and if it's specified more than once
the values must be the same.

For multi-channel items (HOA), the single-channel rule applies for each
channel, and the value determined for each channel must be the same.

This means that in the typical HOA structure with one pack per order
(pointing to the channels for that order and a pack for the next order
down), the absoluteDistance must be specified in the top-level pack, and
if it's specified in the lower packs it must match.