pack allocation: fix with multiple silent packs
fixes #53

The specific bug is that whenever the result would contain multiple
distinct packs which only contain silent channels, multiple equivalent
solutions would be produced. This could cause an "Ambiguous format
references" error when the references from an audioObject are not in
fact ambiguous.

The data that triggers this is not useful, and causes an error
(rather than incorrect behaviour), so it was likely never triggered in
the wild.

When allocating new packs for silent tracks, the packs need to be
allocated in order, to avoid allocating both [p1, p2] and [p2, p1] with
silent tracks, as in the original bug. This is implemented by modifying
the list of possible packs to allocate in subsequent steps, by creating
remaining_packs in candidate_new_packs, and threading that through.