avs: fix structure validation
Previously, validation only worked if there was one audioContent in an
audioProgramme with alternativeValueSets.

This also changes the intended behaviour, so that an audioProgramme and
a contained audioContent can not reference the same alternativeValueSet.
This is done because if an audioContent references an
alternativeValueSet, there is no need to reference it from the
audioProgramme.

This turned into a general refactor of this validation, because the old
code had the wrong structure to be able to work properly.